Skip resource.cfg if NEMS 1.0

+++ restore.sh
@@ -74,12 +74,16 @@ else
 							 exit
 					 fi
 					 
-					 if [[ -e "/tmp/nems_migrator_restore/etc/nagios3/resource.cfg" ]]; then
-						 rm /etc/nagios3/resource.cfg
-						 cp -p /tmp/nems_migrator_restore/etc/nagios3/resource.cfg /etc/nagios3/
-				   else 
+					 if [[ $backupver == "1.0" ]]; then
+					 	echo "Upgrading to newer version of NEMS: Please edit /etc/nagios3/resource.cfg to configure your mail settings."
+					 else
+						 if [[ -e "/tmp/nems_migrator_restore/etc/nagios3/resource.cfg" ]]; then
+							 rm /etc/nagios3/resource.cfg
+							 cp -p /tmp/nems_migrator_restore/etc/nagios3/resource.cfg /etc/nagios3/
+						  else 
 							 echo "Nagios Configuration Missing. This is a critical error."
 							 exit
+						 fi
 					 fi
